+++ tox.ini
@@ -7,6 +7,7 @@ suppress-none-returning = True
 # docstring-convention = google
 per-file-ignores =
     tests/*:D103,D100
+    backend/app/db.py:F401
 
 [coverage:run]
 omit = tests/*
